requirements.txt
# flask
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
blinker==1.9.0
click==8.1.8
flask==3.1.0
itsdangerous==2.2.0

# db
PyMySQL[rsa]==1.1.1
sqlalchemy==2.0.40
Flask-SQLAlchemy==3.1.1

# additional
faker==37.1.0
flask_login==0.6.3
ruff==0.12.0
python-dotenv==1.1.0
gunicorn==23.0.0
pytest==8.3.5
pytest-mock==3.14.0

# for lab6
alembic==1.13.1
Flask-Migrate==4.0.7

#exam
markdown==3.8.2
bleach==6.2.0
